refactor: create execute func for operation manager
