+++ src/modules/commands/fs/rm/index.js
@@ -1,8 +1,8 @@
 import { rm as remove } from 'fs/promises';
 import { resolve } from 'path';
 
-import { appData } from '../../appData.js';
-import { errorHandler } from '../../errorHandler.js';
+import { appData } from '../../../appData.js';
+import { errorHandler } from '../../../errorHandler.js';
 
 export const rm = async (pathToFile) => {
     const path = resolve(appData.cwd, pathToFile);
