+++ src/modules/commands/os/index.js
@@ -1,3 +1,5 @@
-import { manager } from './manager.js';
+import { operationManager } from './operationManager.js';
 
-export const os = command => manager[command]();
+import { execute } from '../../../utils/execute.js';
+
+export const os = command => execute(operationManager[command]);
