+++ src/modules/init.js
@@ -2,10 +2,10 @@ import { createInterface } from 'readline/promises';
 import { stdin as input, stdout as output } from 'process';
 
 import { appData } from './appData.js';
-import { errorHandler } from './errorHandler.js';
 
 import { operationManager } from './commands/operationManager.js';
 
+import { execute } from '../utils/execute.js';
 import { printMessage } from '../utils/printMessage.js';
 import { parseCommandLine } from '../utils/parseCommandLine.js';
 
@@ -23,9 +23,8 @@ export const init = () => {
         if (line === EXIT_COMMAND) readlineInterface.close();
 
         const [ command, argLine ] = parseCommandLine(line);
-        const operation = operationManager[command] || errorHandler.invalidInput;
 
-        await operation(argLine);
+        await execute(operationManager[command], argLine);
 
         printMessage(appData.messages.cwd);
     });
